Autocompile

+++ programowanie/lab1/zadanie2.c
@@ -5,10 +5,10 @@ int main(void)
 {
     printf("Input coefficients (a b c) of the quadratic equation ax^2 + bx + c = 0. Separate them with spaces.\n");
 
-    float a, b, c;
-    scanf("%f %f %f", &a, &b, &c);
+    double a, b, c;
+    scanf("%lf %lf %lf", &a, &b, &c);
 
-    float delta = b * b - 4 * a * c;
+    double delta = b * b - 4 * a * c;
 
     if (delta < 0)
     {
@@ -16,13 +16,13 @@ int main(void)
     }
     else if (delta == 0)
     {
-        float root = -b / (2 * a);
+        double root = -b / (2 * a);
         printf("One real root: %f\n", root);
     }
     else
     {
-        float root1 = (-b - sqrtf(delta)) / (2 * a);
-        float root2 = (-b + sqrtf(delta)) / (2 * a);
+        double root1 = (-b - sqrtf(delta)) / (2 * a);
+        double root2 = (-b + sqrtf(delta)) / (2 * a);
         printf("Two real roots: %f and %f\n", root1, root2);
     }
